requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
greenlet==3.1.1
humanize==4.12.1
hvac==2.3.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
paramiko==3.5.1
pycparser==2.22
pyfiglet==1.0.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scp==0.15.0
SQLAlchemy==2.0.38
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
